Fix error installing older version
Would be more of an issue if specific versioning was supported (yet?!) -
but I ran into it anyway with changing/removing `device_codename` and
`com.spotify.music`. Rolling out earlier to Pixel devices or something?

(The newer version was also non-functional, lacking some parts of the
split APK, so not really sure what's going on, hopefully that won't
happen when it updates for me without the Pixel (sailfish) codename...)